moved controllers to log controller
