+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/app_start/RouteConfig.kt
@@ -34,6 +34,7 @@ object WebRouteConfig : RouteConfig
             WebAuthRouteConfig.endpoints +
                     WebDocumentRouteConfig.endpoints +
                     WebReportRouteConfig.endpoints +
+                    WebLogsRouteConfig.endpoints +
                     WebVersionRouteConfig.endpoints +
                     WebUserRouteConfig.endpoints +
                     WebPermissionRouteConfig.endpoints +
