+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/controllers/web/ReportController.kt
@@ -136,16 +136,16 @@ class ReportController(
         return okayResponse()
     }
 
-    fun running(): List<Running>
-    {
-        // val profile = context.userProfile
+    // fun running(): List<Running>
+    // {
+    //     // val profile = context.userProfile
 
-        // val emailAddress = profile!!.id
-        val user = context.userProfile!!.id
-        // val user = context.userProfile?.id
-        // val user = "test.user@example.com"
-        return reportRunRepository.getAllRunningReports(user)
-    }
+    //     // val emailAddress = profile!!.id
+    //     val user = context.userProfile!!.id
+    //     // val user = context.userProfile?.id
+    //     // val user = "test.user@example.com"
+    //     return reportRunRepository.getAllRunningReports(user)
+    // }
 
     fun publishReports(): String
     {
